Refactor: Remove now-unnecessary button grouping logic
